requirements.txt
DrissionPage==4.1.0.9
colorama==0.4.6
python-dotenv
pyinstaller
requests
aiohttp==3.9.4
click==8.1.7
certifi==2024.2.2